fix: Kamek LR-continuation hook discovery and dispatch
